The sword works enough. See description.
It's janky; I hate it. But it works. enough. The changes I would make if there were maybe 5 more of me would be to clean up the file and fine tune, but I've had enough of this game from implementing this feature alone, so I'm gonna throw in the towel.